protobuf refactoring + cascading services refactoring and update

+++ token/services/identity/x509/crypto/configbuilder.go
@@ -19,7 +19,7 @@ const (
 	KeyStoreDirName  = "keystore"
 	PrivSKFileName   = "priv_sk"
 
-	ProtobufProtocolVersionV1 uint64 = 1
+	ProtobufProtocolVersionV1 uint32 = 1
 )
 
 func LoadConfig(dir string, keyStoreDirName string) (*Config, error) {
